bug fix: allow jest to exit successfully
Jest will not exit until all asynchronous operations are finished. When
running the test suite, one is presented with the following error:

    Jest did not exit one second after the test run has completed.

In order to allow jest to exit cleanly, the database connection pool
maintained by knex needs to be destroyed. Add an afterAll() handler to
setupTests.ts which calls a function provided in src/api/db.ts for this
purpose.