api:submissions: add concept of "comments"
Add a comments model to the API where comments are associated with
submissions. Add DB functions to create and retrieve comments along with
a new route to get comments for an individual submission and create a
new comment.

Only the original submitter or a journalist can add comments to a
subscription.